\name{vegandocs-package}
\alias{vegandocs-package}
\alias{vegandocs}
\docType{package}
\title{
\packageTitle{vegandocs}
}
\description{
\packageDescription{vegandocs}
}
\details{
  The sole purpose of the package is to produce vignettes of tutorials
  with the current version of \pkg{vegan}
}
\author{
\packageAuthor{vegandocs}

Maintainer: \packageMaintainer{vegandocs}
}

\keyword{ package }

